Extract `SQLite3::Quoting` to appropriate file
Like https://github.com/rails/rails/blob/master/activerecord/lib/active_record/connection_adapters/postgresql/quoting.rb
