+++ activerecord/lib/active_record/connection_adapters/sqlite3/quoting.rb
@@ -1,0 +1,45 @@
+module ActiveRecord
+  module ConnectionAdapters
+    module SQLite3
+      module Quoting # :nodoc:
+        def quote_string(s)
+          @connection.class.quote(s)
+        end
+
+        def quote_table_name_for_assignment(table, attr)
+          quote_column_name(attr)
+        end
+
+        def quote_column_name(name)
+          @quoted_column_names[name] ||= %Q("#{name.to_s.gsub('"', '""')}")
+        end
+
+        private
+
+        def _quote(value)
+          case value
+          when Type::Binary::Data
+            "x'#{value.hex}'"
+          else
+            super
+          end
+        end
+
+        def _type_cast(value)
+          case value
+          when BigDecimal
+            value.to_f
+          when String
+            if value.encoding == Encoding::ASCII_8BIT
+              super(value.encode(Encoding::UTF_8))
+            else
+              super
+            end
+          else
+            super
+          end
+        end
+      end
+    end
+  end
+end
